Add .env to the .gitignore & add dotenv package
The dotenv package handles parsing the .env file which contains environment variables that will be read in production. The database connection details should be private and thus will be stored locally in the .env file and not in the repo. On the production server, the environment variables will be passes to the Docker container through the -e command.